Add updateFile action and update file reducer
This commit adds an action to fetch a file from the server. It adds
simple actions to request and recieve the file and a thunk action to
combine the two.

It also updates the reducer to handle the actions.

It also adds a selector to get a file by it's ID.